+++ app/Http/Controllers/SectionsController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\AddSection;
 use App\Http\Requests\UpdateSection;
 use App\Section;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class SectionsController extends Controller
 {
@@ -90,9 +91,21 @@ class SectionsController extends Controller
         if (!$section) {
             $response['message'] = 'Section does not exist';
             return $response;
+        } else {
+            // Check if the given section contains students
+            $students = DB::table('students')
+                ->select('students.id')
+                ->where('section_id', $id)
+                ->get();
+            $studentsArr = json_decode(json_encode($students), true);
+            if (!$studentsArr) {
+                $section->delete();
+                $response['message'] = 'Section deleted successfully';
+                return $response;
+            } else {
+                $response['message'] = 'Can not delete section that contains students';
+                return $response;
+            }
         }
-        $section->delete();
-        $response['message'] = 'Section deleted successfully';
-        return $response;
     }
 }
